cleanup: adopt cobra rootCmd
This is a more standard style.  For now, it makes our version setting
a bit more messy, but we are likely to refactor that during this soon
anyhow.
